Version 0.30
List of new features and bug fixes:
- Background images are now closed at the end of Clean tool.
- Added "Display weighted images?" checkbox in the analyse parameters
(located in Param Tool only). Allows to display the ratiometric image
ponderated with the intensity value of the donor.
- Deleted "Save ROIs?" from the settings, they are now always saved
(overwriting if it existed).
- Uses Java windows instead of Fiji windows for dialogs asking to choose
a file or directory on Mac OS X.11 and later in order for them to
display their title on Mac OS X.11 and for getDirectory to work properly
on Mac OS X.12.
- Outputs a Dialog instead of an error if TurboReg plugin is not
installed. The Dialog tells where to find the plugin and how to install
it.
- Fixed last image of a stack not being normalized.
- Plot ratio vs intensity (ratio) now puts current Ratio image in front
with current region. It also saves path to intensity images and ROIs and
try to open them if they exist.